+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/api
 
+## 1.0.0-rc.35
+
+### Patch Changes
+
+- Updated dependencies [[`3b7af80`](https://github.com/thorswap/SwapKit/commit/3b7af8085ecc7e279dbd639b3375a81cc740970a)]:
+  - @swapkit/types@1.0.0-rc.34
+
 ## 1.0.0-rc.34
 
 ### Patch Changes