+++ packages/toolboxes/evm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @swapkit/toolbox-evm
 
+## 1.0.0-rc.72
+
+### Patch Changes
+
+- Updated dependencies [[`692e678`](https://github.com/thorswap/SwapKit/commit/692e678dbad40165c133ba4d2db2d97f5dd22283)]:
+  - @swapkit/helpers@1.0.0-rc.66
+
+## 1.0.0-rc.71
+
+### Patch Changes
+
+- Updated dependencies [[`3b7af80`](https://github.com/thorswap/SwapKit/commit/3b7af8085ecc7e279dbd639b3375a81cc740970a)]:
+  - @swapkit/helpers@1.0.0-rc.65
+  - @swapkit/types@1.0.0-rc.34
+
 ## 1.0.0-rc.70
 
 ### Patch Changes