+++ packages/wallets/wc/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @swapkit/wallet-wc
 
+## 1.0.0-rc.81
+
+### Patch Changes
+
+- Updated dependencies [[`692e678`](https://github.com/thorswap/SwapKit/commit/692e678dbad40165c133ba4d2db2d97f5dd22283), [`9c363d7`](https://github.com/thorswap/SwapKit/commit/9c363d7c6403e0dd6f1cef00cb3f79620c6e4a62)]:
+  - @swapkit/helpers@1.0.0-rc.66
+  - @swapkit/toolbox-cosmos@1.0.0-rc.74
+  - @swapkit/toolbox-evm@1.0.0-rc.72
+
+## 1.0.0-rc.80
+
+### Patch Changes
+
+- Updated dependencies [[`3b7af80`](https://github.com/thorswap/SwapKit/commit/3b7af8085ecc7e279dbd639b3375a81cc740970a)]:
+  - @swapkit/helpers@1.0.0-rc.65
+  - @swapkit/types@1.0.0-rc.34
+  - @swapkit/toolbox-cosmos@1.0.0-rc.73
+  - @swapkit/toolbox-evm@1.0.0-rc.71
+
 ## 1.0.0-rc.79
 
 ### Patch Changes